Move test/catch.hpp to include/lib/catch.hpp
Issue #32
