+++ test/catch.cpp
@@ -1,2 +1,1 @@
-#define CATCH_CONFIG_MAIN  // This tells Catch to provide a main() - only do this in one cpp file
-#include "catch.hpp"
+#include "lib/catch.hpp"
